% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{brodmann}
\alias{brodmann}
\title{Brodmann Areas Atlas}
\usage{
brodmann()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the Brodmann areas cortical parcellation
with 39 regions per hemisphere. Contains both 2D polygon geometry
for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
brodmann()
}
\references{
Brodmann K (1909). Vergleichende Lokalisationslehre der
Grosshirnrinde. Leipzig: Johann Ambrosius Barth.
Pijnenburg R, et al. (2021). \emph{NeuroImage}, 239, 118274.
\doi{10.1016/j.neuroimage.2021.118274}
}
\concept{ggseg_atlases}
