% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brodmann}
\alias{brodmann}
\alias{brodmann_3d}
\title{Brodmann atlas}
\format{
An object of class \code{brain_atlas} of length 4.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
brodmann

brodmann_3d
}
\description{
A historical Brodmann atlas based on files
released with  Pijnenburg et al., NeuroImage, 
239, 2021, (\href{https://doi.org/10.1016/j.neuroimage.2021.118274}{DOI})
Version 1; 15-01-2021.
\itemize{
 \item{brodmann}{ - brodmann atlas}
 \item{brodmann_3d}{ - brodmann 3d mesh atlas}
}
}
\examples{
data(brodmann)
data(brodmann_3d)
}
\references{
Brodmann K (1909). "Vergleichende Lokalisationslehre der Grosshirnrinde". Leipzig: Johann Ambrosius Barth.
    (\href{https://digital.zbmed.de/zbmed/id/554966?}{ZB med})
}
\concept{ggseg_atlases ggseg3d_atlases}
\keyword{datasets}
